+++ changelog.md
@@ -14,6 +14,7 @@
 - added declare argument to declare cpp code after str manipulation
 - saved cli command in cmgrdf_cli_command env var
 - added is_in_command util to perform regex matching on the cli command
+- added fullTraceback debug option
 
 fix:
 - Added safeguards against empty regions (0 events)